requirements.txt
alembic==1.13.1
beanie==1.24.0
fastapi==0.108.0
pip-chill==1.0.3
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
sqlalchemy-utils==0.41.1
uvicorn==0.25.0
